+++ src/gradient/Gradient.ts
@@ -9,7 +9,7 @@ import { matrixToSVG } from '../util/misc/svgParsing';
 import { linearDefaultCoords, radialDefaultCoords } from './constants';
 import { parseColorStops } from './parser/parseColorStops';
 import { parseCoords } from './parser/parseCoords';
-import { parseType, parseGradientUnits } from './parser/misc';
+import { parseType } from './parser/misc';
 import type {
   ColorStop,
   GradientCoords,
@@ -20,6 +20,7 @@ import type {
 } from './typedefs';
 import { classRegistry } from '../ClassRegistry';
 import { isPath } from '../util/typeAssertions';
+import type { SVGParsingOptions } from '../parser/elements_parser';
 
 /**
  * Gradient class
@@ -380,37 +381,37 @@ export class Gradient<
    */
   static fromElement(
     el: SVGGradientElement,
-    instance: FabricObject,
-    svgOptions: SVGOptions
+    {
+      viewBoxWidth,
+      viewBoxHeight,
+      width,
+      height,
+      gradientUnits,
+      opacity,
+      offsetX,
+      offsetY,
+    }: SVGParsingOptions & SVGOptions
   ): Gradient<GradientType> {
-    const gradientUnits = parseGradientUnits(el);
-    const center = instance._findCenterFromElement();
     return new this({
       id: el.getAttribute('id') || undefined,
       type: parseType(el),
       coords: parseCoords(el, {
-        width: svgOptions.viewBoxWidth || svgOptions.width,
-        height: svgOptions.viewBoxHeight || svgOptions.height,
+        width: viewBoxWidth || width || 0,
+        height: viewBoxHeight || height || 0,
       }),
-      colorStops: parseColorStops(el, svgOptions.opacity),
+      colorStops: parseColorStops(el, opacity),
       gradientUnits,
       gradientTransform: parseTransformAttribute(
         el.getAttribute('gradientTransform') || ''
       ),
-      ...(gradientUnits === 'pixels'
-        ? {
-            offsetX: instance.width / 2 - center.x,
-            offsetY: instance.height / 2 - center.y,
-          }
-        : {
-            offsetX: 0,
-            offsetY: 0,
-          }),
+      offsetX,
+      offsetY,
     });
   }
   /* _FROM_SVG_END_ */
 }
 
 classRegistry.setClass(Gradient, 'gradient');
+classRegistry.setSVGClass(Gradient, 'gradient');
 classRegistry.setClass(Gradient, 'linear');
 classRegistry.setClass(Gradient, 'radial');